# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: PGFPlotsGenerator API
  description: '智能图表生成系统后端接口。除标注外均需 `Authorization: Bearer <token>`。'
  version: "1.0"
  contact:
    name: PGFPlotsGenerator
servers:
  - url: http://localhost:3000
    description: 本地默认后端地址（见 src/config.js）
...
security:
  - BearerAuth: []
paths:
...
  /api/history:
    get:
      tags: [历史]
      summary: 历史列表（分页/搜索/日期过滤）
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string, description: 匹配描述或数据集名 }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      records:
                        type: array
                        items: { $ref: "#/components/schemas/HistoryRecord" }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      total_pages: { type: integer }
        "400":
          description: 分页参数无效
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/history/{id}:
    get:
      tags: [历史]
      summary: 历史详情
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/HistoryRecord" }
        "404":
          description: 不存在或无权访问
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
    delete:
      tags: [历史]
      summary: 删除历史（事务内级联删除 api_log）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
        "404":
          description: 不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: '用户/管理员登录返回的 JWT（请求头 `Authorization: Bearer <token>`）'
  schemas:
...
    MessageSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
    MessageFail:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
...
    HistoryRecord:
      type: object
      description: 历史记录条目（历史页返回字段）
      properties:
        id: { type: integer }
        history_id: { type: integer }
        user_id: { type: integer }
        data_id: { type: integer, nullable: true }
        description: { type: string }
        chart_code: { type: string }
        generation_path: { type: string, nullable: true }
        file_name: { type: string }
        created_at: { type: string }
...